feat(Output): Output message enum 추가
